GasLiftStage2: avoid implicit narrowing

+++ opm/simulators/wells/GasLiftStage2.cpp
@@ -390,7 +390,7 @@ mpiSyncLocalToGlobalGradVector_(const std::vector<GradPair>& grads_local,
 
     std::vector<int> sizes_(this->comm_.size());
     std::vector<int> displ_(this->comm_.size() + 1, 0);
-    int mySize = grads_local_tmp.size();
+    int mySize = static_cast<int>(grads_local_tmp.size());
     this->comm_.allgather(&mySize, 1, sizes_.data());
     std::partial_sum(sizes_.begin(), sizes_.end(), displ_.begin()+1);
     std::vector<Pair> grads_global_tmp(displ_.back());
